updated the strcat program 12

+++ 0x06-pointers_arrays_strings/0-strcat.c
@@ -10,8 +10,8 @@
  */
 char *_strcat(char *dest, char *src)
 {
-size_t i = 0;
-size_t j = 0;
+int i = 0;
+int j = 0;
 
 while (dest[i] != '\0')
 {
